Initial adapter (#13)
* removed integration tests and old adapter

* Add sample and refdoc
