+++ spring-cloud-function-adapters/spring-cloud-function-adapter-gcp/pom.xml
@@ -2,12 +2,17 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<artifactId>spring-cloud-function-adapter-gcp</artifactId>
+	<name>Spring Cloud Function Adapter GCP</name>
+	<description>Spring Cloud Function Adapter for Google Cloud Functions</description>
+
 	<parent>
 		<artifactId>spring-cloud-function-adapter-parent</artifactId>
 		<groupId>org.springframework.cloud</groupId>
 		<version>3.0.4.BUILD-SNAPSHOT</version>
 	</parent>
-	<modelVersion>4.0.0</modelVersion>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
@@ -16,8 +21,6 @@
 		<gson.version>2.8.5</gson.version>
 	</properties>
 
-
-	<artifactId>spring-cloud-function-adapter-gcp</artifactId>
 	<dependencies>
 		<dependency>
 			<groupId>com.google.cloud.functions</groupId>
@@ -49,12 +52,6 @@
 			<groupId>com.google.cloud.functions.invoker</groupId>
 			<artifactId>java-function-invoker</artifactId>
 			<version>${google.cloud.functions.version}</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.awaitility</groupId>
-			<artifactId>awaitility</artifactId>
-			<version>3.1.6</version>
 			<scope>test</scope>
 		</dependency>
 	</dependencies>
