+++ spring-cloud-function-samples/function-sample-gcp/.gitignore
@@ -1,0 +1,1 @@
+deploy/
