+++ spring-cloud-function-samples/function-sample-gcp/src/main/resources/META-INF/MANIFEST.MF
@@ -1,0 +1,1 @@
+Main-Class: com.example.CloudFunctionMain
